Reshuffle some code
