+++ components/ratgdo/secplus1.h
@@ -83,7 +83,7 @@ namespace secplus1 {
 
         Result call(Args args);
 
-        const ProtocolTraits& traits() const { return this->traits_; }
+        const Traits& traits() const { return this->traits_; }
 
     protected:
         void wall_panel_emulation(size_t index = 0);
@@ -129,7 +129,7 @@ namespace secplus1 {
         uint32_t last_tx_ { 0 };
         uint32_t last_status_query_ { 0 };
 
-        ProtocolTraits traits_;
+        Traits traits_;
 
         SoftwareSerial sw_serial_;
